Float cast and directions method

+++ libs/dodulad_ci/ci.hpp
@@ -37,7 +37,7 @@ namespace ci {
 
 	template <typename F> void iter(F& f1, F& f2);
 
-	__device__ void cudaIter(node_calc* ncData, size_t ncSize, double* f1, double* f2);
+	__device__ void cudaIter(node_calc* ncData, size_t ncSize, float* f1, float* f2);
 
 	void finalize();
